% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_highlight.R
\name{annotate_highlights}
\alias{annotate_highlights}
\title{Annotate Highlightes With Predicates}
\usage{
annotate_highlights(.predicate, label, grouped = TRUE, ...)
}
\description{
Annotate Highlightes With Predicates
}
