% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_highlight.R
\name{geom_highlight}
\alias{geom_highlight}
\title{Clone Other Layers}
\usage{
geom_highlight(..., n = NULL, unhighlighted_colour = ggplot2::alpha("grey",
  0.7), max_highlight = 5L)
}
\arguments{
\item{...}{Expressions to filter data, which is passed to \code{\link[dplyr:filter]{dplyr::filter()}}.}

\item{n}{Number of layers to clone.}

\item{unhighlighted_colour}{Colour for unhighlited lines/points.}

\item{max_highlight}{Max number of series to highlight.}
}
\description{
Clone Other Layers
}
