% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_line.R
\name{geom_highlighted_line}
\alias{geom_highlighted_line}
\title{Highlight Lines With Predicate}
\usage{
geom_highlighted_line(.predicate, mapping = NULL,
  unhighlighted_colour = ggplot2::alpha("grey", 0.3), ...)
}
\description{
Highlight Lines With Predicate
}
