% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghighlight-package.R
\docType{package}
\name{gghighlight-package}
\alias{gghighlight-package}
\title{Highlight Lines and Points in 'ggplot2'}
\description{
Make it easier to explore data with highlights.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yutannihilation.github.io/gghighlight/}
  \item \url{https://github.com/yutannihilation/gghighlight}
  \item Report bugs at \url{https://github.com/yutannihilation/gghighlight/issues}
}

}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com} (\href{https://orcid.org/0000-0002-3385-7233}{ORCID})

}
