% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghighlight-package.R, R/gghighlight.R
\docType{package}
\name{gghighlight}
\alias{gghighlight}
\alias{gghighlight-package}
\title{Highlight Points And Lines In 'ggplot2'}
\usage{
gghighlight(data, mapping, predicate,
  unhighlighted_colour = ggplot2::alpha("grey", 0.3), geom = c("line",
  "point"), ..., environment = parent.frame())
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[=fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{predicate}{expression to filter data, which is passed to \code{\link[dplyr:filter]{dplyr::filter()}}.}

\item{unhighlighted_colour}{colour for unhighlited lines/points}

\item{geom}{name of geometry}

\item{...}{arguments passed to the corresponding geometry functions (e.g. \code{geom_line()})}

\item{environment}{If a variable defined in the aesthetic mapping is not
found in the data, ggplot will look for it in this environment. It defaults
to using the environment in which \code{ggplot()} is called.}
}
\description{
Make it easier to explore data with highlights
}
