% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghighlights-package.R
\docType{package}
\name{gghighlights}
\alias{gghighlights}
\alias{gghighlights-package}
\title{Highlight Points And Lines In 'ggplot2'}
\description{
Make it easier to explore data with highlights
}
