% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_highlight.R
\name{scale_highlight}
\alias{scale_highlight}
\alias{scale_highlight_colour}
\alias{scale_highlight}
\alias{scale_highlight_color}
\alias{scale_highlight}
\alias{scale_highlight_fill}
\title{Highlight With Predicates}
\usage{
scale_highlight_colour(.predicate, ..., .default_colour = NULL)

scale_highlight_color(.predicate, ..., .default_colour = NULL)

scale_highlight_fill(.predicate, ..., .default_colour = NULL)
}
\arguments{
\item{.predicate}{an expression to be used as a predicate for highlight}

\item{...}{other argument passed to \code{highlight_scale()}}

\item{.default_colour}{colour for unhighlighted elements}
}
\description{
Highlight With Predicates
}
