% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_highlight.R
\name{scale_highlight_colour}
\alias{scale_highlight_colour}
\title{Highlight With Predicates}
\usage{
scale_highlight_colour(.predicate, ...)
}
\arguments{
\item{.predicate}{an expression to be used as a predicate for highlight}

\item{...}{other argument passed to \code{highlight_scale()}}
}
\description{
Highlight With Predicates
}
