% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_highlight.R
\name{geom_highlight}
\alias{geom_highlight}
\title{Clone Other Layers}
\usage{
geom_highlight(..., n = NULL, unhighlighted_colour = ggplot2::alpha("grey",
  0.7))
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{n}{Number of layers to clone.}
}
\description{
Clone Other Layers
}
