% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghighlight-old.R
\name{gghighlight-old}
\alias{gghighlight-old}
\alias{gghighlight_line}
\alias{gghighlight_point}
\title{Highlight Data With Predicate}
\usage{
gghighlight_line(...)

gghighlight_point()
}
\arguments{
\item{...}{Ignored.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("defunct")}

\code{gghiglight_line()} and \code{gghighlight_point()} are deprecated. Please use \code{\link[=gghighlight]{gghighlight()}} instead.
}
\keyword{internal}
