% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghighlight.R
\name{gghighlight}
\alias{gghighlight}
\alias{gghighlight_line}
\title{Highlight Data With Predicate}
\usage{
gghighlight(data, mapping, predicate_quo,
  unhighlighted_colour = ggplot2::alpha("grey", 0.3),
  geom_func = ggplot2::geom_blank, group_key = NULL, ...,
  environment = parent.frame())

gghighlight_line(data, mapping, predicate,
  unhighlighted_colour = ggplot2::alpha("grey", 0.3), use_group_by = TRUE,
  ..., environment = parent.frame())
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[=fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{unhighlighted_colour}{Colour for unhighlited lines/points.}

\item{...}{Arguments passed to the corresponding geometry functions (e.g. \code{geom_line()}).}

\item{environment}{If a variable defined in the aesthetic mapping is not
found in the data, ggplot will look for it in this environment. It defaults
to using the environment in which \code{ggplot()} is called.}

\item{predicate}{Expression to filter data, which is passed to \code{\link[dplyr:filter]{dplyr::filter()}}.}

\item{use_group_by}{If \code{TRUE}, apply \code{filter} on the grouped data.}
}
\description{
Highlight Data With Predicate
}
\examples{
d <- data.frame(
  idx = c( 1, 1, 1, 2, 2, 2, 3, 3, 3),
  value = c( 1, 2, 3,10,11,12, 9,10,11),
  category = rep(c("a","b","c"), 3),
  stringsAsFactors = FALSE
)

gghighlight_line(d, aes(idx, value, colour = category),
                 max(value) > 10)

}
