% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_highlight.R
\name{scale_highlight_colour}
\alias{scale_highlight_colour}
\title{Highlight With Predicates}
\usage{
scale_highlight_colour(.predicate, ..., type = "seq", palette = 1,
  direction = 1)
}
\description{
Highlight With Predicates
}
